Update User model to match latest server model (#113)
* Flatten "custom" blob on User model and clean up.

User fields "name", "location", "about", and "avatar" have been moved
out of a "custom" dictionary and into the top level.

They are all optional, and should be handled as truly optional and not
forced to empty strings at the model level. If a certain view needs to
treat a nil value as an empty string, it can do that itself.

User model had two different serializations which are now combined.

* Include avatarPath in User serialization

The server doesn't need or use it - but it's reasonable to include for
completeness.

Show the avatar in the edit screen as well.