Fixes: Clean up new feature regressions (#147)
* Server look-up paymentAddress in Pay flow

Because it is now possible to message with someone who is not a contact,
and because we used the local contacts database to look up their cached
paymentAddress, the Pay flow would fail when talking to a non-contact
as no paymentAddress was available.

Accepting a payment request was unaffected, because then the
paymentAddress is available in the payment request payload.

* Resize avatar image before storing/uploading

On typical consumer broadband, images were taking ~20 seconds to upload,
during which the UI remains frozen. By resizing to a lower resolution
and dismissing the cropping view while the upload is still in progress,
we are closer to the needed experience.

Ideally we will also use JPEG representations rather than PNG.

* Clear RequiresSignIn key and dismiss sign in

We were only clearing RequiresSignIn after a phrase restore, but not
after creating a new user. This caused a user to be permenantly locked
out of the app after tapping 'create new account' from the sign in
screen.

We now clear RequiresSignIn and dismiss sign in view in all scenarios
where the app is transitioning into having a valid User.current