Use standard push notifications, build 44
Due to Apple restrictions, we cannot use voip push notifications unless
we include voice calling features. For now we are reverting to standard
push notifications, and losing the ability to show the sender and
message contents as a result of not being able to reliably wake up the
app in the backround for message decryption.